producer/requirements.txt
kafka-python
faker
requests
